refactor: タグ検索の固定定義を分離
